initial diffdrivetrain setup
